+++ R/save_piece_obj.R
@@ -8,8 +8,8 @@
 #' @param res Resolution of the faces.
 #' @return A list with named elements "obj", "mtl", "png" with the created filenames.
 #' @examples
-#'     cfg <- game_systems("dejavu3d")$piecepack
-#'     files <- save_piece_obj("tile_face", suit = 3, rank = 3, cfg = cfg)
+#'     cfg <- game_systems("sans3d")$dominoes
+#'     files <- save_piece_obj("tile_face", suit = 3+1, rank=6+1, cfg = cfg)
 #'     print(files)
 #' @seealso See \code{\link{geometry_utils}} for a discussion of the 3D rotation parameterization.
 #' @export